requirements.txt
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
ciscoconfparse==1.5.19
colorama==0.4.3
cryptography==3.0
dnspython==2.0.0
future==0.18.2
idna==2.10
Jinja2==2.11.2
junos-eznc==2.2.1
lxml==4.5.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
napalm==2.5.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.2.0
nornir==2.4.0
nxapi-plumbing==0.5.2
paramiko==2.7.1
passlib==1.7.2
pycparser==2.20
pyeapi==0.8.3
pyIOSXR==0.53
PyNaCl==1.4.0
pyserial==3.4
python-decouple==3.3
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scp==0.13.2
six==1.15.0
textfsm==1.1.0
typing-extensions==3.7.4.2
urllib3==1.25.10
